requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.1; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
assemblyai==0.20.1; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.1.0; python_version >= '3.7'
beautifulsoup4==4.12.2; python_full_version >= '3.6.0'
bs4==0.0.1
cachecontrol==0.13.1; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.7
curl-cffi==0.5.10; python_version >= '3.7'
dataclasses-json==0.6.3; python_version >= '3.7' and python_version < '4.0'
decorator==4.4.2
distro==1.8.0; python_version >= '3.6'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
docx2txt==0.8
duckduckgo-search==4.1.0; python_version >= '3.8'
exceptiongroup==1.2.0; python_version < '3.11'
faiss-gpu==1.7.2
fastapi==0.105.0; python_version >= '3.8'
ffmpeg==1.4
firebase-admin==6.3.0; python_version >= '3.7'
frozenlist==1.4.1; python_version >= '3.8'
google-api-core[grpc]==2.15.0; python_version >= '3.7'
google-api-python-client==2.111.0; python_version >= '3.7'
google-auth==2.25.2; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-firestore==2.14.0; python_version >= '3.7'
google-cloud-storage==2.14.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.7.0; python_version >= '3.7'
google-search-results==2.4.2; python_version >= '3.5'
googleapis-common-protos==1.62.0; python_version >= '3.7'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx[brotli,http2,socks]==0.26.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
imageio==2.33.1; python_version >= '3.4'
imageio-ffmpeg==0.4.9; python_version >= '3.4'
install==1.3.5; python_version >= '2.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
langchain==0.0.352; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.6; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.3; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.0.74; python_version < '4.0' and python_full_version >= '3.8.1'
loguru==0.7.2; python_version >= '3.5'
lxml==4.9.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
marshmallow==3.20.1; python_version >= '3.8'
more-itertools==7.0.0; python_version >= '3.4'
moviepy==1.0.3
msgpack==1.0.7; python_version >= '3.8'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.2; python_version != '2.7'
openai==1.6.1; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
pinecone-client==2.2.4; python_version >= '3.8'
proglog==0.1.10
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.1; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pygame==2.5.2; python_version >= '3.6'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pyparsing==3.1.1; python_version >= '3.1'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==1.1.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
pytube==15.0.0; python_version >= '3.7'
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==69.0.2; python_version >= '3.8'
simplify-docx==0.1.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-sdk==3.26.1; python_full_version >= '3.6.0'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
tenacity==8.2.3; python_version >= '3.7'
tiktoken==0.5.2; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
typing-inspect==0.9.0
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.1.0; python_version >= '3.8'
uvicorn==0.25.0; python_version >= '3.8'
websockets==12.0; python_version >= '3.8'
wikipedia==1.4.0
wincertstore==0.2
yarl==1.9.4; python_version >= '3.7'
youtube-dl==2021.12.17
youtube-transcript-api==0.6.1
